backend/requirements.txt
aiopg==1.4.0
alembic==1.12.0
asyncpg==0.28.0
bandit==1.7.5
black==23.9.0
databases==0.8.0
fastapi==0.103.0
flake8==6.1.0
hatch==1.7.0
mypy==1.5.1
pip==22.3.1
pipdeptree==2.13.0
pydantic-settings==2.0.3
setuptools==65.5.0
uvicorn==0.23.2
